requirements.txt
requests==2.25.1
PyJWT==2.6.0
PyYAML==6.0
